testsitearea-main/requirements.txt
Flask
Flask-Cors
google-cloud-speech
gunicorn